The transmitter widget now shows 1 for amp on (instead of 14)
I suspect that the original coder for this "widget" just copied and pasted part of the code, from where he is calculating the color-grading for transmit power which he used to paint the GAIN / AMP texts, into the label showing the amp status itself.

With the AMP turned on, his calculation adds 14 to the value used to measure the red-ness of the text color. So the same line of code was pasted when populating the text showing for the amp status.

But the convention in all other places inside portapack is that amp=1 means turned on, 0 means off.