nightly update on scanner
This update includes a working manual scan tab.

Also, it works OK if no stored TXT frequencies are available (or no microSD is inserted): It will just start up showing the MANUAL SCAN tab. If you go to the STORED tab, it will show a "NO SCANNER TXT FOUND".

THERE IS A NEW BUTTON, labeled as "AUDIO APP" below, which is NOT WORKING: It CRASHES the portapack, when pressed.

Hopefully tomorrow it will work, and allow you to jump right into the RX -> AUDIO app to analyze the found frequency. It is intended as a convenient shortcut.