New "nightly update" for scanner
Now, if db reaches its MAX capacity, it states so.

On scan description, it shows the frequencies in better format.

There is an AUDIO APP button which loads the Analog Audio App with the current freq and parameters, for better analysis / fine tune / recording.

Lastly, but not least, I took off the spaces at the beginning of the frequency shown by function to_string_short_freq