"nightly update"
This update has everything working. The scanner is not ready, nevertheless, because I realized that the whole TAB system is not necessary: Everything fits in just one screen, thus probably simplifying code and interface.

I expect to upload a further update, with the "non TABS scanner" some hours from now, if this new vision cristalizes correctly.