Scanner Enhancement
when the scanner detects a signal power above the "squelch" setting, instead of inmediately stopping and activating the audio, it locks into the frequency, listening for some extra time (less than a second) and confirms that the signal power is sustained (and not something spureous), before pausing the scan and  enabling the user to listen.

The big frequency numbers now change color based on the above procedure:

Grey numbers while scanning.
Yellow numbers when locking into freq
Green numbers when freq is confirmed and scan paused, enabling audio.