Added RX to MIC TX (2way radio)
Still Work in Progress!

The vumeter is acting up, particularly when starting / stopping the RX Audio, it will dissappear (but you can get it back by "flash" transmitting).

Also I activated the audio peak indicator on the vumeter, so you can better see how your MIC is responding to your voice.

Lastly, I changed the colors on the vumeter: It's grey, when not energized by a signal. Maybe it is somewhat less "shiny" but under my eyes, it is far more effective as an audio vumeter.