MIC TX RX and a FILEMAN BUG squashed
MIC TX: new code, trying to fix the vumeter which misteriously dissappears momentarily when activating / RX audio. (it reappears when TXing)

Touch: Some fine-tuning, nevermind the value, it works fine

FILEMAN: fixed the RENAME BUG (it renamed the file but path was missing from new name, so the file ended on sd root).