Improvements over MIC TX RX and Radiosonde apps
1) MIC TX RX: I replicated some of the enhancements discussed in this PR:

https://github.com/eried/portapack-mayhem/pull/161

including:

* Separate frequency configuration for RX and TX
* LNA VGA and AMP on/off for RX
* GAIN and AMP on/off for TX

I left out the PTT "PUSH TO TALK SCREEN BUTTON" because I am not sure at this time it will NOT interfere with other apps, due to some inner modifications done for it to work, inside the widget library.

2) Added subframe decoding for the Vaisala radiosonde, including TEMP and HUMIDITY, as also added the TIMESTAMP (YYYYMMMDDHHMMSS) in which the last packet was received, as a reference.

There are other minor changes, most notably the function that returns the decimal value with rounding capabilities was moved from the whipcalc app, into the more general library "string_format" so it can be used by other apps (it is currently being used also by the radiosonde app, in order to give a rounded up value on temp /humidity for Vaisala).