ongoing work on OOK app and De Bruijn
Debug version, very slow. There are problems with the de-bruijn algorithm included by Furrtek: it seems to have two bugs.

1) When called with compute for the first time, it returns an extra bit, which carries on subsequent returns.

2) The results are wrong by 1 bit, apparently, periodically (maybe related to 1) bug.

Also the encoders definition ... what is the encoder 16-bits and why it has a repetition AFSK value of 50 ?

Need some expert eyes on this code, hopefully someone with expertise can review it.